backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
numpy>=1.24
praat-parselmouth>=0.4.4
imageio-ffmpeg>=0.4.9
